feat(gateway): introduce module protocol
