+++ app/gateway/src/lib.rs
@@ -1,17 +1,12 @@
 //! Walrus gateway — application shell composing runtime, channels,
 //! sessions, authentication, and cron scheduling.
 
-pub mod builder;
 pub mod channel;
 pub mod config;
 mod feature;
-pub mod hook;
-pub mod session;
-pub mod state;
+pub mod protocol;
 pub mod utils;
-pub mod ws;
 
-pub use builder::build_runtime;
 pub use channel::{
     auth::{AuthContext, AuthError, Authenticator},
     key::ApiKeyAuthenticator,
@@ -22,6 +17,4 @@ pub use feature::{
     cron::{CronJob, CronScheduler},
     memory::MemoryBackend,
 };
-pub use hook::GatewayHook;
-pub use session::{Session, SessionManager, SessionScope, TrustLevel};
-pub use state::AppState;
+pub use protocol::{Gateway, GatewayHook, builder::build_runtime, session::SessionManager};
