+++ app/gateway/src/protocol/builder.rs
@@ -2,7 +2,7 @@
 
 use crate::MemoryBackend;
 use crate::config::MemoryBackendKind;
-use crate::hook::GatewayHook;
+use crate::protocol::GatewayHook;
 use anyhow::Result;
 use deepseek::DeepSeek;
 use llm::LLM;
